Movement and Camera Work.
More work on the movement, I think it's almost perfect now. Also setup
the camera to move angles depending on the character's height, seems
good for now, we'll see how practical it is. Also made a Fan block that
pushes you up, though the smoothness of the motion could be better.